Typo

+++ scanner/engine/evaluate_worker.cpp
@@ -129,7 +129,7 @@ bool PreEvaluateWorker::yield(i32 item_size,
   entry.needs_reset = first_item_ ? needs_reset_ : false;
   entry.last_in_io_packet = (r + item_size >= total_rows_) ? true : false;
   entry.warmup_rows =
-      std::max(std::min(item_size, work_entry.warmup_rows - r), 0);
+      std::max(std::min(item_size, (i32)(work_entry.warmup_rows - r)), 0);
   entry.columns.resize(work_entry.columns.size());
   entry.last_in_task = work_entry.last_in_task;
